+++ java/core/src/main/java/eu/mais_h/mathsync/FullContent.java
@@ -14,15 +14,15 @@ import org.json.JSONTokener;
  * Summary implementation simply using items content.
  */
 public class FullContent implements Summary {
-  
+
   static final Summary EMPTY = new FullContent(Collections.<EquatableArray>emptySet(), Collections.<EquatableArray>emptySet());
-  
+
   private final Set<EquatableArray> added;
   private final Set<EquatableArray> removed;
-  
+
   /**
    * Deserializes a summary from its JSON representation.
-   * 
+   *
    * @param tokener the JSON content.
    */
   FullContent(JSONTokener tokener) {
@@ -30,7 +30,7 @@ public class FullContent implements Summary {
     added = deserializeArray(deserialized.getJSONArray("added"));
     removed = deserializeArray(deserialized.getJSONArray("removed"));
   }
-  
+
   private FullContent(Set<EquatableArray> added, Set<EquatableArray> removed) {
     this.added = added;
     this.removed = removed;
@@ -50,6 +50,24 @@ public class FullContent implements Summary {
     Set<EquatableArray> removedCopy = new HashSet<>(removed);
     while (items.hasNext()) {
       insertOrRemove(addedCopy, removedCopy, items.next());
+    }
+    return new FullContent(addedCopy, removedCopy);
+  }
+
+  @Override
+  public Summary minus(byte[] item) {
+    Set<EquatableArray> addedCopy = new HashSet<>(added);
+    Set<EquatableArray> removedCopy = new HashSet<>(removed);
+    insertOrRemove(removedCopy, addedCopy, item);
+    return new FullContent(addedCopy, removedCopy);
+  }
+
+  @Override
+  public Summary minus(Iterator<byte[]> items) {
+    Set<EquatableArray> addedCopy = new HashSet<>(added);
+    Set<EquatableArray> removedCopy = new HashSet<>(removed);
+    while (items.hasNext()) {
+      insertOrRemove(removedCopy, addedCopy, items.next());
     }
     return new FullContent(addedCopy, removedCopy);
   }
@@ -83,7 +101,7 @@ public class FullContent implements Summary {
   public Difference<byte[]> toDifference() {
     return new SerializedDifference(unwrap(added), unwrap(removed));
   }
-  
+
   private void insertOrRemove(Set<EquatableArray> mayInsert, Set<EquatableArray> mayRemove, byte[] item) {
     EquatableArray wrapped = new EquatableArray(item);
     if (mayRemove.contains(wrapped)) {
@@ -92,7 +110,7 @@ public class FullContent implements Summary {
       mayInsert.add(wrapped);
     }
   }
-  
+
   private JSONArray serializeSet(Set<EquatableArray> set) {
     JSONArray array = new JSONArray();
     for (EquatableArray ea : set) {
@@ -100,7 +118,7 @@ public class FullContent implements Summary {
     }
     return array;
   }
-  
+
   private Set<byte[]> unwrap(Set<EquatableArray> wrapped) {
     Set<byte[]> unwrapped = new HashSet<>();
     for (EquatableArray ea : wrapped) {
@@ -108,7 +126,7 @@ public class FullContent implements Summary {
     }
     return unwrapped;
   }
-  
+
   private Set<EquatableArray> deserializeArray(JSONArray array) {
     Set<EquatableArray> deserialized = new HashSet<>();
     for (int i = 0; i < array.length(); i++) {
@@ -116,9 +134,9 @@ public class FullContent implements Summary {
     }
     return deserialized;
   }
-  
+
   private static final class EquatableArray {
-    
+
     private final byte[] content;
     private final int hashCode;
 
@@ -130,7 +148,7 @@ public class FullContent implements Summary {
     private EquatableArray(String serialized) {
       this(Defaults.deserialize(serialized));
     }
-    
+
     @Override
     public int hashCode() {
       return hashCode;
@@ -150,7 +168,7 @@ public class FullContent implements Summary {
       }
       return true;
     }
-    
+
     @Override
     public String toString() {
       return Defaults.serialize(content);