+++ doc/src/changelog.md
@@ -3,9 +3,14 @@ layout: default
 title: MathSync
 ---
 
+# On the road to 0.5.0
+
+* remove deprecated method `Summary#minus(Summary)` and deprecated class `ResolverFromSummarizers`
+
 # On the road to 0.4.0
 
 * remove small summary generation from larger ones (concept is too weak)
+* deprecate general summary substraction (concept is too weak - will be replaced on a case by case basis)
 * rationalize javascript packages into a single one
 
 # 0.3.0